\name{Dataset}
\alias{.Random.seed}
\alias{tpod}
\alias{gen}
\alias{chr}
\alias{fam}
\alias{y}
\title{
  Tetra-seed Pods
}
\description{
Two biparental crosses phenotyped for the percentage of pods containing four seeds
}
\usage{
  data(tpod)
}
\details{
Soybean nested association panel with 2 families (\eqn{fam}) containing 196 indiviuals. Genotypic matrix (\eqn{gen}) have 376 SNP across 20 chromosome (\eqn{chr}). Phenotypic information (\eqn{y}) regards the proportion of tetra-seed pods. Data provided by Rainey Lab for Soybean Breeding and Genetics, Purdue University.
}
\author{
  Alencar Xavier and Katy Rainey
}