\name{bWGR-package}
\alias{bWGR-package}
\alias{bWGR}
\docType{package}
\title{
Bagging Whole-Genome Regression
}
\description{
Designed for whole-genome regressions using Bayesian models fitted via EM or Gibbs sampling with optional resampling techniques (Bagging) and polygenic term with reduced dimensionality.
}
\details{
\tabular{ll}{
Package: \tab bWGR\cr
Type: \tab Package\cr
Version: \tab 1.3.2 \cr
Date: \tab 2016-09-19\cr
License: \tab GPL-3 \cr
}
}
\author{
Alencar Xavier, Shizhong Xu, William Muir, Katy Rainey.

Maintainer: Alencar Xavier <alenxav@gmail.com>
}