\name{bWGR-package}
\alias{bWGR-package}
\alias{bWGR}
\docType{package}
\title{
Bagging Whole-Genome Regression
}
\description{
Whole-genome regression methods on Bayesian framework fitted via EM
or Gibbs sampling, with optional sampling techniques and kernel term.
}
\details{
\tabular{ll}{
Package: \tab bWGR\cr
Type: \tab Package\cr
Version: \tab 1.4 \cr
Date: \tab 2017-03-21\cr
License: \tab GPL-3 \cr
}
}
\author{
Alencar Xavier, Shizhong Xu, William Muir, Katy Rainey.

Maintainer: Alencar Xavier <alenxav@gmail.com>
}