\name{bWGR-package}
\alias{bWGR-package}
\alias{bWGR}
\docType{package}
\title{
Bagging Whole-Genome Regression
}
\description{
Designed for whole-genome regressions using Bayesian models fitted via Gibbs sampling with optional resampling techniques (Bagging and Bootstrapping) and polygenic term with reduced dimensionality.
}
\details{
\tabular{ll}{
Package: \tab bWGR\cr
Type: \tab Package\cr
Version: \tab 1.2 \cr
Date: \tab 2015-10-28\cr
License: \tab GPL-3 \cr
}
}
\author{
Alencar Xavier, William Muir, Katy Rainey.

Maintainer: Alencar Xavier <xaviera@purdue.edu>
}